+++ lib/urls.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require_relative 'urls/validator'
+require_relative 'urls/shortener'
+
+# URLs related functionality
+module URLs
+end