+++ spec/spec_helper.rb
@@ -12,11 +12,15 @@ Dir[Config.root.join('spec/support/**/*.rb')].each { |f| require f }
 
 RSpec.configure do |config|
   config.include Rack::Test::Methods, type: :api
+  config.include AppMixin, type: :api
 
   config.profile_examples = true
   config.order = :random
   config.expect_with :rspec do |expectations|
     expectations.include_chain_clauses_in_custom_matcher_descriptions = true
+    # For some reason Rspec doesn't recognize CustomErrors::* as a specific
+    # error class
+    expectations.on_potential_false_positives = :nothing
   end
 
   config.mock_with :rspec do |mocks|