+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Changelog
 =========
 
 ## Unreleased
-
+- Added Url shortener module to handle creation of ShortUrls
+- Added Validator module to handle validations of short_codes for ShortUrls
 ## [0.1.1] - 2017-10-21
 - Added model documentation
 - Added ShortUrl model